Moved all history related text constants to service.ServiceConstants class
