+++ api/api-job-instance-v1/src/test/groovy/com/thoughtworks/go/apiv1/jobinstance/representers/JobInstancesRepresenterTest.groovy
@@ -49,7 +49,7 @@ class JobInstancesRepresenterTest {
           "state"                : "Building",
           "result"               : "Unknown",
           "original_job_id"      : null,
-          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime() + "",
+          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime(),
           "rerun"                : false,
           "agent_uuid"           : "1234",
           "pipeline_name"        : "pipeline",
@@ -101,7 +101,7 @@ class JobInstancesRepresenterTest {
           "state"                : "Building",
           "result"               : "Unknown",
           "original_job_id"      : null,
-          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime() + "",
+          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime(),
           "rerun"                : false,
           "agent_uuid"           : "1234",
           "pipeline_name"        : "pipeline",
@@ -139,7 +139,7 @@ class JobInstancesRepresenterTest {
           "state"                : "Building",
           "result"               : "Unknown",
           "original_job_id"      : null,
-          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime() + "",
+          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime(),
           "rerun"                : false,
           "agent_uuid"           : "1234",
           "pipeline_name"        : "pipeline",
@@ -172,7 +172,7 @@ class JobInstancesRepresenterTest {
           "state"                : "Building",
           "result"               : "Unknown",
           "original_job_id"      : null,
-          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime() + "",
+          "scheduled_date"       : jobInstances.get(0).getScheduledDate().getTime(),
           "rerun"                : false,
           "agent_uuid"           : "1234",
           "pipeline_name"        : "pipeline",
