+++ server/src/main/java/com/thoughtworks/go/server/service/JobInstanceService.java
@@ -52,7 +52,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.thoughtworks.go.server.service.PipelineHistoryService.BAD_CURSOR_MSG;
+import static com.thoughtworks.go.server.service.ServiceConstants.History.BAD_CURSOR_MSG;
 import static java.lang.String.format;
 
 @Service
