+++ server/src/test-fast/java/com/thoughtworks/go/server/service/JobInstanceServiceTest.java
@@ -529,11 +529,11 @@ public class JobInstanceServiceTest {
 
             assertThatCode(() -> jobService.getJobHistoryViaCursor(username, pipelineName, stageName, jobConfigName, -10, 0, 10))
                     .isInstanceOf(BadRequestException.class)
-                    .hasMessage("The query parameter `after`, if specified, must be a positive integer.");
+                    .hasMessage("The query parameter 'after', if specified, must be a positive integer.");
 
             assertThatCode(() -> jobService.getJobHistoryViaCursor(username, pipelineName, stageName, jobConfigName, 0, -10, 10))
                     .isInstanceOf(BadRequestException.class)
-                    .hasMessage("The query parameter `before`, if specified, must be a positive integer.");
+                    .hasMessage("The query parameter 'before', if specified, must be a positive integer.");
         }
     }
