+++ server/src/test-fast/java/com/thoughtworks/go/server/service/StageServiceTest.java
@@ -648,11 +648,11 @@ public class StageServiceTest {
 
             assertThatCode(() -> stageService.findStageHistoryViaCursor(username, pipelineName, STAGE_NAME, -10, 0, 10))
                     .isInstanceOf(BadRequestException.class)
-                    .hasMessage("The query parameter `after`, if specified, must be a positive integer.");
+                    .hasMessage("The query parameter 'after', if specified, must be a positive integer.");
 
             assertThatCode(() -> stageService.findStageHistoryViaCursor(username, pipelineName, STAGE_NAME, 0, -10, 10))
                     .isInstanceOf(BadRequestException.class)
-                    .hasMessage("The query parameter `before`, if specified, must be a positive integer.");
+                    .hasMessage("The query parameter 'before', if specified, must be a positive integer.");
         }
     }
